backend/requirements.txt
grpcio
grpcio-tools
python-dotenv
pyyaml
pyroute2
python-dotenv
fastapi[all]
pony
SQLAlchemy
fastapi_utils
ipaddr
sqlalchemy_utils
pyjwt
passlib